+++ HelpTechService/Attention/Interfaces/REST/Resources/Job/JobResource.cs
@@ -4,6 +4,6 @@
         (int Id, int AgendaId, int ConsumerId,
         DateTime? AnswerDate, DateTime? WorkDate,
         string Address, string Description,
-        decimal? Time, double? LaborBudget,
-        double? MaterialBudget, string JobState);
+        decimal? Time, decimal? LaborBudget,
+        decimal? MaterialBudget, string JobState);
 }
